# ctrlc/public/swagger.yaml
openapi: 3.0.0
info:
  title: My API
  description: API Documentation with Swagger UI
  version: 1.0.0
paths:
...
  /api/available_mt5:
    post:
      summary: Get available MT5 subscriber accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  example: "678373fd3bdb581a425a8b9f"
              required:
                - user_id
      responses:
        '200':
          description: List of MT5 accounts with no active subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    mt5_name:
                      type: string
                      example: "John MT5"
                    mt5_id:
                      type: string
                      example: "123456"
                    account_type:
                      type: string
                      example: "subscriber"
        '400':
          description: Missing token or user ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Missing token or user ID"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Server error"
...
  /api/create-token:
    post:
      summary: Save token and verify against MT5 accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  example: "mcjg71kWFn4vMArVmBfPNOtlZMBIDTEm"
                user_id:
                  type: string
                  example: "678373fd3bdb581a425a8b9f"
              required:
                - token
                - user_id
      responses:
        '200':
          description: Token verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: boolean
                    example: true
        '400':
          description: Missing token or user ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Missing token or user ID"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Server error"
  /api/mt5account:
    post:
      summary: Retrieve MT5 accounts for a specific user by user ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  example: "678373fd3bdb581a425a8b9f"
              required:
                - user_id
      responses:
        '200':
          description: MT5 account details successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    mt5_name:
                      type: string
                      example: "John Doe"
                    mt5_id:
                      type: string
                      example: "mt5-123456"
                    account_type:
                      type: string
                      example: "subscriber"
        '400':
          description: Missing user ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Missing token or user ID"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Server error"
  /api/subscribe:
    post:
      summary: Subscribe a Mt5 account to a strategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mt5_id:
                  type: string
                  example: "mt5-123456"
                strategie_id:
                  type: string
                  example: "64dbd10f8fa8a933dcfc8db1"
              required:
                - mt5_id
                - strategie_id
      responses:
        '200':
          description: Subscription to strategy was successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Subscribe Strategy success"
        '400':
          description: Missing parameters (mt5_id or strategie_id)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Missing params"
        '500':
          description: Server error occurred
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Server error"
  /api/subscribed:
    post:
      summary: Get subscriptions for a user, including strategy and MT5 account information.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  example: "64dbd10f8fa8a933dcfc8db1"
              required:
                - user_id
      responses:
        '200':
          description: Successfully fetched user subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    mt5_name:
                      type: string
                      example: "MT5 Account 123"
                    mt5_id:
                      type: string
                      example: "mt5-123456"
                    account_type:
                      type: string
                      example: "demo"
                    strategy_name:
                      type: string
                      example: "Strategy A"
                    symbol:
                      type: string
                      example: "EURUSD"
                    status:
                      type: string
                      example: "active"
                    sub_id:
                      type: string
                      example: "sub-id-123"
                    start_date:
                      type: string
                      format: date-time
                      example: "2025-04-05T12:00:00Z"
                    end_date:
                      type: string
                      format: date-time
                      example: "2025-06-05T12:00:00Z"
        '400':
          description: Missing user ID parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Missing user ID"
        '500':
          description: Server error occurred
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Server error"
...
  /api/strategyinfo:
    post:
      summary: Fetch a strategy by its name.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: The name of the strategy.
                  example: "EMA Strategy"
              required:
                - name
      responses:
        '200':
          description: Strategy details successfully fetched.
          content:
            application/json:
              schema:
                type: object
                properties:
                  strategy:
                    type: object
                    example: { "_id": "64dbd10f8fa8a933dcfc8db2", "name": "EMA Strategy" }
        '400':
          description: Missing strategy name.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Missing token or user ID"
        '500':
          description: Server error occurred.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "An error occurred while fetching strategies."
  /api/mt5info:
    post:
      summary: Fetch an MT5 account by its ID.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  description: The MT5 account ID.
                  example: "12345"
              required:
                - id
      responses:
        '200':
          description: MT5 account details successfully fetched.
          content:
            application/json:
              schema:
                type: object
                properties:
                  mt5_account:
                    type: object
                    example: { "_id": "64dbd10f8fa8a933dcfc8db2", "mt5_name": "Account 1", "mt5_id": "12345" }
        '400':
          description: Missing MT5 account ID.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Missing ID"
        '500':
          description: Server error occurred.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "An error occurred while fetching strategies."
...